+++ Data/PizzaRepository.cs
@@ -7,8 +7,8 @@ namespace TetePizza.Data
 {
     public class PizzaRepository : IPizzaRepository
     {
-        private  List<Pizza> Pizzas { get; set; }
-        private  int nextIdIngredient = 3;
+        private List<Pizza> Pizzas { get; set; }
+        private int nextIdIngredient = 3;
 
         public PizzaRepository()
         {
@@ -29,7 +29,7 @@ namespace TetePizza.Data
                             Calories = "Cada Tomate contiene dos 22 calorías.",
                             ExpiryDate = "20/02/2024.",
                             Origin = "El tomate proviene de Andalucía",
-                            Price = 0.22,
+                            Price =(decimal)0.22,
                             NutritionalInfo = "Consumir tomates es beneficioso para la salud debido a la amplia variedad de nutrientes y compuestos bioactivos que ofrecen.",
                             IsGlutenFree = true
                         },
@@ -42,7 +42,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 200 calorías por cada 100 gramos de prosciutto.",
                             ExpiryDate = "2024-02-22",
                             Origin = "Italia",
-                            Price = 1.3,
+                            Price = (decimal)1.3,
                             NutritionalInfo = "Rico en grasas saludables y proteínas.",
                             IsGlutenFree = true
                         },
@@ -55,7 +55,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 110 calorías por cada 28 gramos de queso Parmesano.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.5,
+                            Price = (decimal)2.5,
                             NutritionalInfo = "Alto contenido de calcio y proteínas.",
                             IsGlutenFree = true
                         },
@@ -68,12 +68,12 @@ namespace TetePizza.Data
                             Calories = "No tiene calorías.",
                             ExpiryDate = "2024-03-01",
                             Origin = "Región Mediterránea",
-                            Price = 3.0,
+                            Price = (decimal)3.0,
                             NutritionalInfo = "Rico en grasas saludables.",
                             IsGlutenFree = true
                         },
                     },
-                    Price = 7.02
+                    Price = (decimal)7.02
                 },
                 new Pizza
                 {
@@ -90,7 +90,7 @@ namespace TetePizza.Data
                             Calories = "Cada tomate contiene aproximadamente 22 calorías.",
                             ExpiryDate = "20/02/2024",
                             Origin = "Andalucía",
-                            Price = 0.22,
+                            Price = (decimal)0.22,
                             NutritionalInfo = "Beneficioso para la salud debido a la variedad de nutrientes y compuestos bioactivos que ofrece.",
                             IsGlutenFree = true
                         },
@@ -103,7 +103,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 7 calorías por cada taza de espinacas.",
                             ExpiryDate = "2024-02-18",
                             Origin = "Local",
-                            Price = 0.3,
+                            Price = (decimal)0.3,
                             NutritionalInfo = "Rica en vitaminas y minerales, baja en calorías.",
                             IsGlutenFree = true
                         },
@@ -116,7 +116,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 11 calorías por cada 100 gramos de champiñones.",
                             ExpiryDate = "2024-02-15",
                             Origin = "Cultivados",
-                            Price = 0.25,
+                            Price = (decimal)0.25,
                             NutritionalInfo = "Buena fuente de proteínas, vitaminas y minerales.",
                             IsGlutenFree = true
                         },
@@ -129,12 +129,12 @@ namespace TetePizza.Data
                             Calories = "No tiene calorías.",
                             ExpiryDate = "2024-03-01",
                             Origin = "Región Mediterránea",
-                            Price = 3.0,
+                            Price = (decimal)3.0,
                             NutritionalInfo = "Rico en grasas saludables.",
                             IsGlutenFree = true
                         },
                     },
-                    Price = 6.77
+                    Price = (decimal)6.77
                 },
                 new Pizza
                 {
@@ -151,7 +151,7 @@ namespace TetePizza.Data
                             Calories = "Cada Tomate contiene dos 22 calorías.",
                             ExpiryDate = "20/02/2024.",
                             Origin = "El tomate proviene de Andalucía",
-                            Price = 0.22,
+                            Price = (decimal)0.22,
                             NutritionalInfo = "Consumir tomates es beneficioso para la salud debido a la amplia variedad de nutrientes y compuestos bioactivos que ofrecen.",
                             IsGlutenFree = true
                         },
@@ -164,7 +164,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 250 calorías por cada 100 gramos de pepperoni.",
                             ExpiryDate = "2024-02-22",
                             Origin = "Estados Unidos",
-                            Price = 1.5,
+                            Price = (decimal)1.5,
                             NutritionalInfo = "Rico en grasas y proteínas, pero consumir con moderación debido a su contenido calórico.",
                             IsGlutenFree = true
                         },
@@ -177,7 +177,7 @@ namespace TetePizza.Data
                             Calories = "Aporta insignificantes calorías.",
                             ExpiryDate = "Indeterminada",
                             Origin = "Mediterráneo",
-                            Price = 0.1,
+                            Price = (decimal)0.1,
                             NutritionalInfo = "Aporta sabor y aroma a la pizza, sin calorías significativas.",
                             IsGlutenFree = true
                         },
@@ -190,7 +190,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 110 calorías por cada 28 gramos de queso Parmesano.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.5,
+                            Price = (decimal)2.5,
                             NutritionalInfo = "Alto contenido de calcio y proteínas.",
                             IsGlutenFree = true
                         },
@@ -203,12 +203,12 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 5 calorías por cada aceituna.",
                             ExpiryDate = "20/02/2024",
                             Origin = "Mediterráneo",
-                            Price = 0.3,
+                            Price = (decimal)0.3,
                             NutritionalInfo = "Ricas en grasas saludables y antioxidantes.",
                             IsGlutenFree = true
                         },
                     },
-                    Price = 9.12
+                    Price = (decimal)9.12
                 },
                 new Pizza
                 {
@@ -225,7 +225,7 @@ namespace TetePizza.Data
                             Calories = "Cada Tomate contiene dos 22 calorías.",
                             ExpiryDate = "20/02/2024.",
                             Origin = "El tomate proviene de Andalucía",
-                            Price = 0.22,
+                            Price = (decimal)0.22,
                             NutritionalInfo = "Consumir tomates es beneficioso para la salud debido a la amplia variedad de nutrientes y compuestos bioactivos que ofrecen.",
                             IsGlutenFree = true
                         },
@@ -238,7 +238,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 80 calorías por cada 28 gramos de queso mozzarella.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.0,
+                            Price = (decimal)2.0,
                             NutritionalInfo = "Bajo contenido de grasa y alto contenido de proteínas y calcio.",
                             IsGlutenFree = true
                         },
@@ -251,7 +251,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 110 calorías por cada 28 gramos de queso cheddar.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Inglaterra",
-                            Price = 1.5,
+                            Price = (decimal)1.5,
                             NutritionalInfo = "Sabor fuerte y color naranja distintivo.",
                             IsGlutenFree = true
                         },
@@ -264,7 +264,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 100 calorías por cada 28 gramos de queso gorgonzola.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.2,
+                            Price = (decimal)2.2,
                             NutritionalInfo = "Queso azul con sabor fuerte.",
                             IsGlutenFree = true
                         },
@@ -277,7 +277,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 80 calorías por cada 28 gramos de queso feta.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Grecia",
-                            Price = 2.0,
+                            Price = (decimal)2.0,
                             NutritionalInfo = "Queso blanco y desmenuzable con sabor salado.",
                             IsGlutenFree = true
                         },
@@ -290,7 +290,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 98 calorías por cada 28 gramos de queso provolone.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.0,
+                            Price = (decimal)2.0,
                             NutritionalInfo = "Queso italiano semiduro con sabor suave.",
                             IsGlutenFree = true
                         },
@@ -303,7 +303,7 @@ namespace TetePizza.Data
                             Calories = "Aproximadamente 110 calorías por cada 28 gramos de queso parmesano.",
                             ExpiryDate = "2024-02-25",
                             Origin = "Italia",
-                            Price = 2.5,
+                            Price = (decimal)2.5,
                             NutritionalInfo = "Alto contenido de calcio y proteínas.",
                             IsGlutenFree = true
                         },
@@ -316,7 +316,7 @@ namespace TetePizza.Data
                             Calories = "Aporta insignificantes calorías.",
                             ExpiryDate = "Indeterminada",
                             Origin = "Mediterráneo",
-                            Price = 0.1,
+                            Price = (decimal)0.1,
                             NutritionalInfo = "Aporta sabor y aroma a la pizza, sin calorías significativas.",
                             IsGlutenFree = true
                         },
@@ -329,23 +329,23 @@ namespace TetePizza.Data
                             Calories = "No tiene calorías.",
                             ExpiryDate = "2024-03-01",
                             Origin = "Región Mediterránea",
-                            Price = 3.0,
+                            Price = (decimal)3.0,
                             NutritionalInfo = "Rico en grasas saludables.",
                             IsGlutenFree = true
                         },
                     },
-                    Price = 10.5
+                    Price = (decimal)10.5
                 },
             };
         }
 
-        public  List<Pizza> GetAll() => Pizzas;
+        public List<Pizza> GetAll() => Pizzas;
 
-        public  Pizza? Get(int IdIngredient) => Pizzas.FirstOrDefault(p => p.Id == IdIngredient);
+        public Pizza? Get(int IdIngredient) => Pizzas.FirstOrDefault(p => p.Id == IdIngredient);
 
         public void Add(Pizza pizza)
         {
-            pizza.Id= nextIdIngredient++;
+            pizza.Id = nextIdIngredient++;
             Pizzas.Add(pizza);
         }
